+++ ios/Classes/LocationPlugin.m
@@ -10,6 +10,7 @@
 @property (strong, nonatomic) CLLocationManager *clLocationManager;
 @property (copy, nonatomic)   FlutterResult      flutterResult;
 @property (assign, nonatomic) BOOL               locationWanted;
+@property (assign, nonatomic) BOOL               permissionWanted;
 
 @property (copy, nonatomic)   FlutterEventSink   flutterEventSink;
 @property (assign, nonatomic) BOOL               flutterListening;
@@ -34,6 +35,7 @@
 
     if (self) {
         self.locationWanted = NO;
+        self.permissionWanted = NO;
         self.flutterListening = NO;
         self.hasInit = NO;
     }
@@ -65,6 +67,11 @@
             };
 
             self.clLocationManager.desiredAccuracy = [dictionary[call.arguments[@"accuracy"]] doubleValue];
+            double distanceFilter = [call.arguments[@"distanceFilter"] doubleValue];
+            if (distanceFilter == 0){
+                distanceFilter = kCLDistanceFilterNone;
+            }
+            self.clLocationManager.distanceFilter = distanceFilter;
             result(@(1));
         }
     } else if ([call.method isEqualToString:@"getLocation"]) {
@@ -95,12 +102,10 @@
             result(@(0));
         }
     } else if ([call.method isEqualToString:@"requestPermission"]) {
+        self.flutterResult = result;
+        self.permissionWanted = YES;
         [self requestPermission];
-        if ([self isPermissionGranted]) {
-            result(@(1));
-        } else {
-            result(@(0));
-        }
+
     } else if ([call.method isEqualToString:@"serviceEnabled"]) {
         if ([CLLocationManager locationServicesEnabled]) {
             result(@(1));
@@ -212,9 +217,18 @@
     if (status == kCLAuthorizationStatusDenied) {
         // The user denied authorization
         NSLog(@"User denied permissions");
+        if (self.permissionWanted) {
+            self.permissionWanted = NO;
+            self.flutterResult(@(0));
+        }
+        
     }
     else if (status == kCLAuthorizationStatusAuthorizedWhenInUse || status == kCLAuthorizationStatusAuthorizedAlways) {
         NSLog(@"User granted permissions");
+        if (self.permissionWanted) {
+            self.permissionWanted = NO;
+            self.flutterResult(@(1));
+        }
 
         if (self.locationWanted || self.flutterListening) {
             [self.clLocationManager startUpdatingLocation];