+++ lib/location.dart
@@ -15,16 +15,8 @@ class LocationData {
   final double heading;
   final double time;
 
-  LocationData._(
-    this.latitude,
-    this.longitude,
-    this.accuracy,
-    this.altitude,
-    this.speed,
-    this.speedAccuracy,
-    this.heading,
-    this.time
-  );
+  LocationData._(this.latitude, this.longitude, this.accuracy, this.altitude,
+      this.speed, this.speedAccuracy, this.heading, this.time);
 
   factory LocationData.fromMap(Map<String, double> dataMap) {
     return LocationData._(
@@ -40,34 +32,27 @@ class LocationData {
   }
 }
 
-/**
- * https://developers.google.com/android/reference/com/google/android/gms/location/LocationRequest
- * https://developer.apple.com/documentation/corelocation/cllocationaccuracy?language=objc
- * Precision of the Location
- */
-enum LocationAccuray { 
-  POWERSAVE,
-  LOW,
-  BALANCED,
-  HIGH,
-  NAVIGATION
-}
-
-
-const MethodChannel _channel = const MethodChannel('lyokone/location');
-const EventChannel _stream = const EventChannel('lyokone/locationstream');
+/// https://developers.google.com/android/reference/com/google/android/gms/location/LocationRequest
+/// https://developer.apple.com/documentation/corelocation/cllocationaccuracy?language=objc
+/// Precision of the Location
+enum LocationAccuracy { POWERSAVE, LOW, BALANCED, HIGH, NAVIGATION }
 
 class Location {
+  static const MethodChannel _channel = const MethodChannel('lyokone/location');
+  static const EventChannel _stream =
+      const EventChannel('lyokone/locationstream');
+
   Stream<LocationData> _onLocationChanged;
 
-  Future<bool> changeSettings({LocationAccuray accuracy = LocationAccuray.HIGH, int interval = 5000}) =>
-    _channel.invokeMethod('changeSettings',
-    {
-      "accuracy": accuracy.index,
-      "interval": interval
-    })
-    .then((result) => result == 1);
-
+  Future<bool> changeSettings(
+          {LocationAccuracy accuracy = LocationAccuracy.HIGH,
+          int interval = 1000,
+          double distanceFilter = 0}) =>
+      _channel.invokeMethod('changeSettings', {
+        "accuracy": accuracy.index,
+        "interval": interval,
+        "distanceFilter": distanceFilter
+      }).then((result) => result == 1);
 
   /// Gets the current location of the user.
   ///
@@ -80,15 +65,17 @@ class Location {
   Future<bool> hasPermission() =>
       _channel.invokeMethod('hasPermission').then((result) => result == 1);
 
+  /// Request the permission to access the location
   Future<bool> requestPermission() =>
-    _channel.invokeMethod('requestPermission').then((result) => result == 1);
+      _channel.invokeMethod('requestPermission').then((result) => result == 1);
 
+  /// Checks if the location service is enabled
   Future<bool> serviceEnabled() =>
-    _channel.invokeMethod('serviceEnabled').then((result) => result == 1);
+      _channel.invokeMethod('serviceEnabled').then((result) => result == 1);
 
+  /// Request the activate of the location service
   Future<bool> requestService() =>
-    _channel.invokeMethod('requestService').then((result) => result == 1);
-
+      _channel.invokeMethod('requestService').then((result) => result == 1);
 
   /// Returns a stream of location information.
   Stream<LocationData> onLocationChanged() {