+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [2.3.3] 31th March 2019
+* Align altitude on Sea Level when available on Android (matching iOS altitude).
+
+## [2.3.2] 27th March 2019
+* Remove GPS limitation on Android
+
+## [2.3.1] 25th March 2019
+* Fixes README
+* Fixes requestPermission not responding the correct result on iOS
+
+## [2.3.0] 22nd March 2019
+* Update example App with proper cancel
+* Add possibility to set accuracy, interval and minimum notification ditance of the requests.
+* Add LocationAccuracy object
+
 ## [2.2.0] 19th March 2019
 * Actually updating locatino when using getLocation (not only relying on LastLocation)
 * Add timestamp to LocationData